Reorganization and renaming.
