+++ Phaxio/ThinRestClient/Entities/Interfaces/IRestClient.cs
