+++ Phaxio/ThinRestClient/Entities/Interfaces/IRestRequest.cs
