+++ Phaxio/ThinRestClient/Entities/Interfaces/IRestResponse.cs
