+++ Phaxio/ThinRestClient/Entities/Method.cs
