+++ Phaxio/ThinRestClient/Entities/Parameter.cs
