+++ Phaxio/ThinRestClient/Entities/RestRequest.cs
