+++ Phaxio/ThinRestClient/Entities/RestResponse.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Phaxio.ThinRestClient.Helpers;
+using System;
 using System.Net;
 using System.Net.Http;
