+++ Phaxio/ThinRestClient/Helpers/HttpClientExtensions.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
 
-namespace Phaxio.ThinRestClient
+namespace Phaxio.ThinRestClient.Helpers
 {
     static class HttpResponseMessageExtensions
     {
